+++ bin/c-mk-restart.js
@@ -1,0 +1,13 @@
+#!/usr/bin/env node
+
+'use strict';
+
+const program = require('commander');
+const { exec } = require('shelljs');
+
+// The basic program, which uses sub-commands.
+program
+    .parse(process.argv);
+
+exec('c mk stop');
+exec('c mk start');